+++ ios/ExpoZebraScannerModule.swift
@@ -26,5 +26,17 @@ public class ExpoZebraScannerModule: Module {
       return "ExpoZebraScanner not available on IOS."
     }
 
+    Function("startCustomScan") { (action: String) -> String in
+      return "ExpoZebraScanner not available on IOS."
+    }
+
+    Function("stopCustomScan") {
+      return "ExpoZebraScanner not available on IOS."
+    }
+
+    AsyncFunction("getDataWedgeVersion") { () -> [Int] in
+      return [0, 0, 0]
+    }
+
   }
 }